Migrated security and parts of request service in manager UI to jsinterop for usage with <or-app> component

+++ manager/client/src/main/java/org/openremote/manager/client/i18n/ManagerMessages.java
@@ -507,7 +507,7 @@ public interface ManagerMessages extends Messages {
 
     String noNotifications();
 
-    String notificationDeleted(Long id);
+    String notificationDeleted(double id);
 
     String deleteNotification();
